Update split_tunnel_VPN.sh
Changed to have it unzip the pia ovpn files into the openvpn folder.
Removed multiple networks, and updated Sweden to the new correct PIA address.
I personally want this to only be run on intial setup, as such I'm assuming I haven't created a user yet, and that my user group for that user will be the same as their name. As such there is only a need to ask for the one variable.